Create codsoft_taskno1
Task: 1. Number Game
This code is for a number-guessing game in which the user tries to guess a randomly generated number between 1 and 100. A random integer between 1 and 100 is produced using a Random object (rand) (rand.nextInt(100) + 1). The code uses a loop to collect user input and feedback. Until the user guesses the number correctly, the software will keep asking them to do so using a while(true) loop. The application congratulates the user and ends the loop if the guessed number and the random number match. It gives a suggestion as to whether the target number is higher or lower than the guess if the guess is wrong. To record the number of attempts, the tryCount is incremented every time the user makes a guess, so the program can display the total number of attempts once the user wins.