Add files via upload
To access the ATM menu, the user must first enter the right PIN, which is 2468 in this case. The program ends if users input the wrong PIN. Every case has the appropriate prompts, balancing adjustments, and is completely encased. Before going back to the menu, the break statements make sure every operation is finished. The user can leave Case 4, which shuts down the scanner and terminates the program in a tidy manner. With PIN security and functional menu options, this code provides the user with organized ATM experience.